Implement :state(foo) alongside :--foo
This patch implements :state(foo) alongsite :--foo so that we can have
both supported on stable at the same time and then deprecate/remove
:--foo. This patch does not enable :state(foo) by default. I will do
that after implementing deprecation for :--foo.

Due to pushback from webkit, the syntax for this feature was discussed
again in the CSSWG, and a resolution was made to go back from :--foo to
:state(foo):
https://github.com/w3c/csswg-drafts/issues/4805#issuecomment-1663111980

This was discussed in an I2S where we decided to only deprecate :--foo
if WebKit ships :--foo, which has now happened.
https://groups.google.com/a/chromium.org/g/blink-dev/c/JvpHoUfhJYE

This patch is a partial revert of http://crrev.com/845429, which renamed
from :state(foo) to :--foo.

Fixed: 1508033
Bug: 1514397
Change-Id: I497fd5acc05ba033e8e72c429568099c89d5d639